+++ src/work/exec/process/_reexport_std.rs
@@ -1,4 +1,4 @@
-// devela::work::process::_reexport_std
+// devela::work::exec::process::_reexport_std
 //
 //!
 //