+++ src/work/exec/process/error.rs
@@ -1,4 +1,4 @@
-// devela::work::process::error
+// devela::work::exec::process::error
 //
 //! Defines [`ExitStatusError`].
 //