+++ src/work/exec/process/mod.rs
@@ -1,7 +1,7 @@
-// devela::work::process
+// devela::work::exec::process
 //
-#![doc = crate::_DOC_WORK_PROCESS!()] // public
-#![doc = crate::_doc!(modules: crate::work; process)]
+#![doc = crate::_DOC_WORK_EXEC_PROCESS!()] // public
+#![doc = crate::_doc!(modules: crate::work::exec; process)]
 #![doc = crate::_doc!(flat:"work")]
 #![doc = crate::_doc!(extends: process)]
 //