+++ src/work/exec/remote/mod.rs
@@ -1,0 +1,19 @@
+// devela::work::exec::remote
+//
+#![doc = crate::_DOC_WORK_EXEC_REMOTE!()] // public
+#![doc = crate::_doc!(modules: crate::work::exec; remote)]
+#![doc = crate::_doc!(flat:"work")]
+#![doc = crate::_doc!(hr)]
+//
+
+// pub mod chain;
+// pub mod rpc;
+
+crate::structural_mods! { // _pub_mods
+    _pub_mods {
+        pub use super::{
+            // chain::_all::*,
+            // rpc::_all::*,
+        };
+    }
+}