+++ src/work/exec/thread/_reexport_std.rs
@@ -1,4 +1,4 @@
-// devela::work::thread::_reexport_std
+// devela::work::exec::thread::_reexport_std
 //
 //!
 //