+++ src/work/plan/mod.rs
@@ -1,0 +1,23 @@
+// devela::work::task
+//
+#![doc = crate::_DOC_WORK_PLAN!()] // public
+#![doc = crate::_doc!(modules: crate::work; plan)]
+#![doc = crate::_doc!(flat:"work")]
+#![doc = crate::_doc!(hr)]
+//
+
+// pub mod budget;
+// pub mod pool;
+// pub mod queue;
+// pub mod sched;
+
+crate::structural_mods! { // _pub_mods
+    _pub_mods {
+        // pub use super::{
+        //     // budget::_all::*,
+        //     // pool::_all::*,
+        //     // queue::_all::*,
+        //     // sched::_all::*,
+        // };
+    }
+}